pyproject.toml
[tool.poetry]
name = "graphql_complexity"
version = "0.3.0"
description = "A python library that provides complexity calculation helpers for GraphQL"
authors = ["Checho3388 <ezequiel.grondona@gmail.com>"]
packages = [
    { include = "graphql_complexity", from = "src" },
]
repository = "https://github.com/Checho3388/graphql-complexity"
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["graphql", "complexity", "query", "mutation", "subscription"]

[tool.poetry.dependencies]
python = "^3.10"
graphql-core = "^3.2.3"
strawberry-graphql = { version = "^0.219.2", optional = true }

[tool.poetry.extras]
strawberry-graphql = ["strawberry-graphql"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
mypy = "^1.8.0"
black = "^24.2.0"
pytest-cov = "^4.1.0"
flake8 = "^7.0.0"

[tool.poetry.group.demo]
optional = true

[tool.poetry.group.demo.dependencies]
strawberry-graphql = { version = "^0.219.2", extras = ["debug-server"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
